Generalize type token to handle functions
